Move lives and Score to singleplayergameui package
<organisation, interface, feature
